requirements.txt
openpyxl>=3.0.0
python-dotenv>=0.19.0
msal>=1.20.0
requests>=2.28.0
